Revert "Chore: Au 340 (#79)"
This reverts commit e1442f789098717340ea23961a3547580fff0917.
